feat: Align editing and layouting steps layouts
