+++ app/configurator/components/configurator.tsx
@@ -1,26 +1,40 @@
-import { Trans } from "@lingui/macro";
-import { Box, SxProps, Typography } from "@mui/material";
+import { t, Trans } from "@lingui/macro";
+import {
+  Box,
+  Grow,
+  SxProps,
+  Tooltip,
+  Typography,
+  useEventCallback,
+} from "@mui/material";
 import Button, { ButtonProps } from "@mui/material/Button";
+import { PUBLISHED_STATE } from "@prisma/client";
+import { useSession } from "next-auth/react";
+import NextLink from "next/link";
 import { useRouter } from "next/router";
 import React, { useEffect, useMemo } from "react";
+import { useClient } from "urql";
+import { useDebounce } from "use-debounce";
 
 import { SelectDatasetStep } from "@/browser/select-dataset-step";
 import { META } from "@/charts";
+import { extractChartConfigComponentIris } from "@/charts/shared/chart-helpers";
 import { ChartPreview } from "@/components/chart-preview";
-import {
-  PublishChartButton,
-  SaveDraftButton,
-} from "@/components/chart-selection-tabs";
 import { HEADER_HEIGHT } from "@/components/header-constants";
 import { Loading } from "@/components/hint";
 import {
-  MetadataPanelStoreContext,
   createMetadataPanelStore,
+  MetadataPanelStoreContext,
 } from "@/components/metadata-panel-store";
+import { useLocalSnack } from "@/components/use-local-snack";
 import {
   ConfiguratorState,
+  enableLayouting,
   getChartConfig,
+  hasChartConfigs,
+  initChartStateFromChartEdit,
   isConfiguring,
+  saveChartLocally,
   useConfiguratorState,
 } from "@/configurator";
 import {
@@ -43,6 +57,9 @@ import {
 } from "@/configurator/components/layout";
 import { LayoutConfigurator } from "@/configurator/components/layout-configurator";
 import { ChartConfiguratorTable } from "@/configurator/table/table-chart-configurator";
+import { useUserConfig } from "@/domain/user-configs";
+import { useDataCubesComponentsQuery } from "@/graphql/hooks";
+import { Icon } from "@/icons";
 import SvgIcChevronLeft from "@/icons/components/IcChevronLeft";
 import { useLocale } from "@/locales/use-locale";
 import { useDataSourceStore } from "@/stores/data-source";
@@ -50,8 +67,11 @@ import {
   InteractiveFiltersChartProvider,
   InteractiveFiltersProvider,
 } from "@/stores/interactive-filters";
+import { createConfig, updateConfig } from "@/utils/chart-config/api";
 import { getRouterChartId } from "@/utils/router/helpers";
+import { replaceLinks } from "@/utils/ui-strings";
 import useEvent from "@/utils/use-event";
+import { useMutate } from "@/utils/use-fetch-data";
 
 const BackContainer = (props: React.PropsWithChildren<{ sx?: SxProps }>) => {
   const { children, sx } = props;
@@ -141,6 +161,227 @@ const BackToMainButton = (props: BackToMainButtonProps) => {
   );
 };
 
+const NextStepButton = (props: React.PropsWithChildren<{}>) => {
+  const { children } = props;
+  const locale = useLocale();
+  const [state, dispatch] = useConfiguratorState(hasChartConfigs);
+  const chartConfig = getChartConfig(state);
+  const componentIris = extractChartConfigComponentIris({ chartConfig });
+  const [{ data: components }] = useDataCubesComponentsQuery({
+    variables: {
+      sourceType: state.dataSource.type,
+      sourceUrl: state.dataSource.url,
+      locale,
+      cubeFilters: chartConfig.cubes.map((cube) => ({
+        iri: cube.iri,
+        componentIris,
+        filters: cube.filters,
+        joinBy: cube.joinBy,
+        loadValues: true,
+      })),
+    },
+  });
+
+  const handleClick = useEvent(() => {
+    if (components?.dataCubesComponents) {
+      dispatch({
+        type: "STEP_NEXT",
+        dataCubesComponents: components.dataCubesComponents,
+      });
+    }
+  });
+
+  return (
+    <Button
+      color="primary"
+      variant="contained"
+      onClick={handleClick}
+      sx={{ minWidth: "fit-content" }}
+    >
+      {children}
+    </Button>
+  );
+};
+
+export const SaveDraftButton = ({
+  chartId,
+}: {
+  chartId: string | undefined;
+}) => {
+  const { data: config, invalidate: invalidateConfig } = useUserConfig(chartId);
+  const session = useSession();
+  const client = useClient();
+
+  const [state, dispatch] = useConfiguratorState();
+
+  const [snack, enqueueSnackbar, dismissSnack] = useLocalSnack();
+  const [debouncedSnack] = useDebounce(snack, 500);
+  const { asPath, replace } = useRouter();
+
+  const createConfigMut = useMutate(createConfig);
+  const updatePublishedStateMut = useMutate(updateConfig);
+  const loggedInId = session.data?.user.id;
+
+  const handleClick = useEventCallback(async () => {
+    try {
+      if (config?.user_id && loggedInId) {
+        const updated = await updatePublishedStateMut.mutate({
+          data: state,
+          published_state: PUBLISHED_STATE.DRAFT,
+          key: config.key,
+        });
+
+        if (updated) {
+          if (asPath !== `/create/${updated.key}`) {
+            replace(`/create/new?edit=${updated.key}`);
+          }
+        } else {
+          throw new Error("Could not update draft");
+        }
+      } else if (state) {
+        const saved = await createConfigMut.mutate({
+          data: state,
+          user_id: loggedInId,
+          published_state: PUBLISHED_STATE.DRAFT,
+        });
+        if (saved) {
+          const config = await initChartStateFromChartEdit(
+            client,
+            saved.key,
+            state.state
+          );
+
+          if (!config) {
+            return;
+          }
+
+          dispatch({ type: "INITIALIZED", value: config });
+          saveChartLocally(saved.key, config);
+          replace(`/create/${saved.key}`, undefined, {
+            shallow: true,
+          });
+        } else {
+          throw new Error("Could not save draft");
+        }
+      }
+
+      enqueueSnackbar({
+        message: (
+          <>
+            {replaceLinks(
+              t({
+                id: "button.save-draft.saved",
+                message: "Draft saved in [My visualisations](/profile)",
+              }),
+              (label, href) => {
+                return (
+                  <div>
+                    <NextLink href={href}>{label}</NextLink>
+                  </div>
+                );
+              }
+            )}
+          </>
+        ),
+        variant: "success",
+      });
+
+      invalidateConfig();
+    } catch (e) {
+      console.log(
+        `Error while saving draft: ${e instanceof Error ? e.message : e}`
+      );
+      enqueueSnackbar({
+        message: t({
+          id: "button.save-draft.error",
+          message: "Could not save draft",
+        }),
+        variant: "error",
+      });
+    }
+
+    setTimeout(() => {
+      updatePublishedStateMut.reset();
+      createConfigMut.reset();
+    }, 2000);
+  });
+
+  const hasUpdated = !!(updatePublishedStateMut.data || createConfigMut.data);
+  const [debouncedHasUpdated] = useDebounce(hasUpdated, 300);
+
+  if (!loggedInId) {
+    return null;
+  }
+
+  return (
+    <Tooltip
+      arrow
+      title={debouncedSnack?.message ?? ""}
+      open={!!snack}
+      disableFocusListener
+      disableHoverListener
+      disableTouchListener
+      onClose={() => dismissSnack()}
+    >
+      <Button
+        endIcon={
+          hasUpdated || debouncedHasUpdated ? (
+            <Grow in={hasUpdated}>
+              <span>
+                <Icon name="check" />
+              </span>
+            </Grow>
+          ) : null
+        }
+        variant="outlined"
+        onClick={handleClick}
+      >
+        <Trans id="button.save-draft">Save draft</Trans>
+      </Button>
+    </Tooltip>
+  );
+};
+
+const LayoutChartButton = () => {
+  return (
+    <NextStepButton>
+      <Trans id="button.layout">Proceed to layout options</Trans>
+    </NextStepButton>
+  );
+};
+
+const PublishChartButton = ({ chartId }: { chartId: string | undefined }) => {
+  const session = useSession();
+  const { data: config } = useUserConfig(chartId);
+  const editingPublishedChart =
+    session.data?.user.id &&
+    config?.user_id === session.data.user.id &&
+    config.published_state === "PUBLISHED";
+
+  return (
+    <NextStepButton>
+      {editingPublishedChart ? (
+        <Box sx={{ display: "flex", alignItems: "center", gap: 2 }}>
+          <Tooltip
+            title={t({
+              id: "button.update.warning",
+              message:
+                "Keep in mind that updating this visualization will affect all the places where it might be already embedded!",
+            })}
+          >
+            <div>
+              <Icon name="hintWarning" />
+            </div>
+          </Tooltip>
+          <Trans id="button.update">Update this visualization</Trans>
+        </Box>
+      ) : (
+        <Trans id="button.publish">Publish</Trans>
+      )}
+    </NextStepButton>
+  );
+};
+
 const ConfigureChartStep = () => {
   const [state, dispatch] = useConfiguratorState();
   const configuring = isConfiguring(state);
@@ -172,9 +413,37 @@ const ConfigureChartStep = () => {
     return null;
   }
 
+  const chartId = getRouterChartId(router.asPath);
+
   return (
     <InteractiveFiltersChartProvider chartConfigKey={chartConfig.key}>
-      <PanelLayout type="LM">
+      <PanelLayout type="LM" sx={{ background: (t) => t.palette.muted.main }}>
+        <PanelHeaderLayout type="LMR">
+          <PanelHeaderWrapper type="L">
+            <BackContainer>
+              <BackButton onClick={handlePrevious}>
+                <Trans id="controls.nav.back-to-preview">Back to preview</Trans>
+              </BackButton>
+            </BackContainer>
+          </PanelHeaderWrapper>
+          <PanelHeaderWrapper
+            type="R"
+            sx={{
+              display: "flex",
+              flexShrink: 0,
+              alignItems: "start",
+              justifyContent: "flex-end",
+              gap: "0.5rem",
+            }}
+          >
+            <SaveDraftButton chartId={chartId} />
+            {enableLayouting(state) ? (
+              <LayoutChartButton />
+            ) : (
+              <PublishChartButton chartId={chartId} />
+            )}
+          </PanelHeaderWrapper>
+        </PanelHeaderLayout>
         <PanelBodyWrapper
           type="L"
           sx={{
@@ -184,11 +453,6 @@ const ConfigureChartStep = () => {
             flexDirection: "column",
           }}
         >
-          <BackContainer>
-            <BackButton onClick={handlePrevious}>
-              <Trans id="controls.nav.back-to-preview">Back to preview</Trans>
-            </BackButton>
-          </BackContainer>
           {chartConfig.chartType === "table" ? (
             <ChartConfiguratorTable state={state} />
           ) : (
